+++ src/main/java/com/github/jlangch/venice/impl/Var.java
@@ -32,29 +32,42 @@ import com.github.jlangch.venice.impl.types.VncVal;
 public class Var implements Serializable {
 
 	public Var(final VncSymbol name, final VncVal val) {
+		this(name, val, true);
+	}
+	
+	public Var(final VncSymbol name, final VncVal val, final boolean overwritable) {
 		this.name = name;
 		this.val = val == null ? Nil : val;
+		this.overwritable = overwritable;
 	}
 	
 	public VncVal getVal() {
 		return val;
 	}
 	
-	public void setVal(final VncVal val) {
-		this.val = val == null ? Nil : val;
-	}
-	
 	public VncSymbol getName() {
 		return name;
 	}
 	
+ 	public boolean isOverwritable() {
+		return overwritable;
+	}
+
 	@Override 
 	public String toString() {
-		return String.format("{%s %s}", name.toString(), val.toString());
+		return String.format(
+				"{%s %s :overwritable %b}", 
+				name.toString(), 
+				val.toString(), 
+				overwritable);
 	}
 	
 	public String toString(final boolean print_readably) {
-		return String.format("{%s %s}", name.toString(print_readably), val.toString(print_readably));
+		return String.format(
+				"{%s %s :overwritable %b}", 
+				name.toString(print_readably), 
+				val.toString(print_readably), 
+				overwritable);
 	}
 	
 
@@ -62,5 +75,6 @@ public class Var implements Serializable {
 	private static final long serialVersionUID = 1598432086227773369L;
 
 	private final VncSymbol name;
-	private VncVal val;
+	private final VncVal val;
+	private final boolean overwritable;
 }