+++ src/main/java/com/github/jlangch/venice/impl/docgen/DocGenerator.java
@@ -866,6 +866,7 @@ public class DocGenerator {
 		all.addSection(generic);
 
 		generic.addItem(getDocItem("def"));
+		generic.addItem(getDocItem("defonce"));
 		generic.addItem(getDocItem("def-dynamic"));
 		generic.addItem(getDocItem("if"));
 		generic.addItem(getDocItem("do"));